refactor: StudyList->StudySecion로 컴포넌트 분리, 조건부 렌더링  단순화
